Update: Put WriteDataSet into own project.
Update: (Again) changed (and tested) namespaces to make source clear and uniform (Data.DataSet.IO and Data.IO.Excel wasn't very uniform (my fault) and it made it a bit more difficult to find the IO that you were searching for).

Update: Put (Excel) Insert operations into own class to make write class less big in the future.
