+++ src/BlazorElectron/BlazorElectron.csproj
@@ -19,7 +19,9 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\DataSetIO\DataSetIO.csproj" />
     <ProjectReference Include="..\DataSet\DataSet.csproj" />
+    <ProjectReference Include="..\ExcelIOComponentTest\DataSetlIOComponentTest.csproj" />
     <ProjectReference Include="..\ExcelIO\ExcelIO.csproj" />
   </ItemGroup>
