+++ src/BlazorElectron/Data/PublicationDataSet/SubClassesModels.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-using Zeiss.PublicationManager.Data.IO.Excel;
+using Zeiss.PublicationManager.Data.Excel.IO;
 using Zeiss.PublicationManager.Data.DataSet;
 
 namespace Zeiss.PublicationManager.Data.DataSet.Model
