+++ src/DataSetIO/DataSetIO.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\DataSet\DataSet.csproj" />
+    <ProjectReference Include="..\ExcelIO\ExcelIO.csproj" />
+  </ItemGroup>
+
+</Project>
